Minor bug fix
Fixed minor bug in the UserResponse model that was preventing the UserResponse from including the user's list of owned/assigned checklist ids in the response body. Also fixed a minor bug that was preventing a user from being deleted due to missing cascading deletes on the users_checklists association table.